refactor: 이름 검증로직을 도메인으로 변경
