+++ src/main/java/racingcar/step4/view/input/InputView.java
@@ -1,4 +1,4 @@
-package racingcar.step4.ui.input;
+package racingcar.step4.view.input;
 
 import racingcar.step4.utils.ConstUtils;
 
@@ -23,11 +23,9 @@ public class InputView {
         System.out.println(helpMessage);
 
         String input = scanner.nextLine();
-        List<String> names = Arrays.stream(input.split(ConstUtils.NAME_SEPARATE_DELIMITER))
+        return Arrays.stream(input.split(ConstUtils.NAME_SEPARATE_DELIMITER))
                 .map(String::trim)
                 .collect(Collectors.toList());
-        validateNames(names);
-        return names;
     }
 
     private int getRoundCountUserInput(String helpMessage) {
@@ -37,16 +35,6 @@ public class InputView {
             return Integer.parseInt(scanner.nextLine());
         } catch (Exception e) {
             throw new NumberFormatException("값은 숫자만 입력할 수 있습니다.");
-        }
-    }
-
-    private void validateNames(List<String> names) {
-        names.forEach(this::validNameLength);
-    }
-
-    private void validNameLength(String name) {
-        if (name.length() > 5) {
-            throw new IllegalArgumentException("이름의 글자는 5자를 초과할 수 없습니다.");
         }
     }
 }
